Moving type definitions around and remove CoreGraphicsGlobalsProvider
