+++ Sources/GlobalsProviders/DefaultGlobalsProvidersSource.swift
@@ -4,7 +4,6 @@ import SwiftRewriterLib
 public class DefaultGlobalsProvidersSource: GlobalsProvidersSource {
     public var globalsProviders: [GlobalsProvider] = [
         CLibGlobalsProviders(),
-        CoreGraphicsGlobalsProvider(),
         UIKitGlobalsProvider(),
         OpenGLESGlobalsProvider(),
         CompoundedMappingTypesGlobalsProvider()
