+++ Tests/GlobalsProvidersTests/DefaultGlobalsProvidersSourceTests.swift
@@ -8,7 +8,6 @@ class DefaultGlobalsProvidersSourceTests: XCTestCase {
         var providers = sut.globalsProviders.makeIterator()
         
         XCTAssert(providers.next() is CLibGlobalsProviders)
-        XCTAssert(providers.next() is CoreGraphicsGlobalsProvider)
         XCTAssert(providers.next() is UIKitGlobalsProvider)
         XCTAssert(providers.next() is OpenGLESGlobalsProvider)
         XCTAssert(providers.next() is CompoundedMappingTypesGlobalsProvider)
